Created Table and Chair in Test Scene
Created chairs and then merged the objects together to make a genuine chair, and not just objects shaped as a chair.

Created a table with the donut thing, and then used a plane to create a floor.

I need to figure out the tools more, and figure out how to make the chairs all at the same height position. All in all I'm surprised I did anything at all.

Saved a second test scene because I was scared it wasn't saving correctly due to a warning in the right hand corner.