Created seat cushions
I realized that I needed some beveled surfaces so I created a square with a soft edge, and placed it on top of the chair. Combined the seat cushion to the chair.

Deleted all the other chairs and replaced them with the cushioned chair.

Next I guess I'll make a new chair (or bench?) with chairs that were made by extruding. Oof