Changed To Bedroom with New Assets
Added a bookshelf by starting with a cube, and a LOT of beveling. Then realized the scene was supposed to be a bedroom(?) so I got rid of 3 chairs, and placed the table and chair in new spots. Created a rug and tried to make it bunch up against the wall.