Got UV Photos of everything. Bumpy Table and Cushy Chair Disappeared
Once again I have no idea what's going on. I checked that every item in the room was cut for UV, and had a color. Then took snapshots of the UV layouts.
When I returned to the room, bumpy table and cushy pillow were missing.
Previous to that, bumpy table wasn't showing the correct colors